Return backup's filesize on backups listing
I am using this plugin in my application and I wanted to find out the size of the backups. I made the change and it is working great for me. I wanted to suggest adding it to your plugin, hence this pull request. Thanks for a great job creating this.